requirements.txt
python-dotenv==1.0.0
psycopg2-binary==2.9.9
mysql-connector-python==8.2.0
pretty_errors==1.2.25
Flask==3.0.0
gunicorn==21.2.0
a2wsgi==1.9.0
daphne==4.0.0
paramiko==3.4.0
sshtunnel==0.4.0
requests==2.31.0
openai==1.9.0